Fix more ifmedia64 fallout in the kernel. It's hiding everywhere...

+++ sys/arch/hppa/gsc/if_ie_gsc.c
@@ -1,4 +1,4 @@
-/*	$OpenBSD: if_ie_gsc.c,v 1.25 2004/10/28 19:13:30 mickey Exp $	*/
+/*	$OpenBSD: if_ie_gsc.c,v 1.26 2014/03/29 18:09:29 guenther Exp $	*/
 
 /*
  * Copyright (c) 1998-2004 Michael Shalayeff
@@ -76,7 +76,7 @@ struct cfattach ie_gsc_ca = {
 	sizeof(struct ie_softc), ie_gsc_probe, ie_gsc_attach
 };
 
-static int ie_gsc_media[] = {
+static uint64_t ie_gsc_media[] = {
 	IFM_ETHER | IFM_10_2,
 };
 #define	IE_NMEDIA	(sizeof(ie_gsc_media) / sizeof(ie_gsc_media[0]))
