+++ sys/arch/sparc/dev/dmavar.h
@@ -1,4 +1,4 @@
-/*	$OpenBSD: dmavar.h,v 1.7 2006/06/02 20:00:54 miod Exp $	*/
+/*	$OpenBSD: dmavar.h,v 1.8 2014/07/28 18:31:39 miod Exp $	*/
 /*	$NetBSD: dmavar.h,v 1.11 1996/11/27 21:49:53 pk Exp $ */
 
 /*
@@ -44,7 +44,7 @@ struct dma_softc {
 	int (*setup)(struct dma_softc *, caddr_t *, size_t *, int, size_t *);
 	void (*go)(struct dma_softc *);
 	u_int	sc_dmactl;
-	int	sc_defaultmedia;		/* default Ethernet media */
+	uint64_t	sc_defaultmedia;	/* default Ethernet media */
 };
 
 #define DMACSR(sc)	(sc->sc_regs->csr)
