+++ sys/arch/sparc/dev/qe.c
@@ -1,4 +1,4 @@
-/*	$OpenBSD: qe.c,v 1.39 2015/03/29 10:59:47 mpi Exp $	*/
+/*	$OpenBSD: qe.c,v 1.40 2015/06/24 09:40:53 mpi Exp $	*/
 
 /*
  * Copyright (c) 1998, 2000 Jason L. Wright.
@@ -864,7 +864,7 @@ qe_ifmedia_upd(ifp)
 	struct ifnet *ifp;
 {
 	struct qesoftc *sc = (struct qesoftc *)ifp->if_softc;
-	int media = sc->sc_ifmedia.ifm_media;
+	uint64_t media = sc->sc_ifmedia.ifm_media;
 
 	if (IFM_TYPE(media) != IFM_ETHER)
 		return (EINVAL);
